Extract index handlers to a separate file
