+++ api/http/handlerfuncs.go
@@ -17,7 +17,6 @@ import (
 	"io"
 	"mime"
 	"net/http"
-	"strings"
 
 	"github.com/go-chi/chi/v5"
 	dshelp "github.com/ipfs/boxo/datastore/dshelp"
@@ -27,7 +26,6 @@ import (
 	"github.com/multiformats/go-multihash"
 	"github.com/pkg/errors"
 
-	"github.com/sourcenetwork/defradb/client"
 	corecrdt "github.com/sourcenetwork/defradb/core/crdt"
 	"github.com/sourcenetwork/defradb/events"
 )
@@ -227,195 +225,6 @@ func patchSchemaHandler(rw http.ResponseWriter, req *http.Request) {
 		simpleDataResponse("result", "success"),
 		http.StatusOK,
 	)
-}
-
-func createIndexHandler(rw http.ResponseWriter, req *http.Request) {
-	db, err := dbFromContext(req.Context())
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	err = req.ParseForm()
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	colNameArg := req.Form.Get("collection")
-	fieldsArg := req.Form.Get("fields")
-	indexNameArg := req.Form.Get("name")
-
-	col, err := db.GetCollectionByName(req.Context(), colNameArg)
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	fields := strings.Split(fieldsArg, ",")
-	fieldDescriptions := make([]client.IndexedFieldDescription, 0, len(fields))
-	for _, field := range fields {
-		fieldDescriptions = append(fieldDescriptions, client.IndexedFieldDescription{Name: field})
-	}
-	indexDesc := client.IndexDescription{
-		Name:   indexNameArg,
-		Fields: fieldDescriptions,
-	}
-	indexDesc, err = col.CreateIndex(req.Context(), indexDesc)
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	type fieldResponse struct {
-		Name      string                `json:"name"`
-		Direction client.IndexDirection `json:"direction"`
-	}
-
-	type indexResponse struct {
-		Name   string          `json:"name"`
-		ID     uint32          `json:"id"`
-		Fields []fieldResponse `json:"fields"`
-	}
-
-	indexResp := indexResponse{
-		Name: indexDesc.Name,
-		ID:   indexDesc.ID,
-	}
-
-	for _, field := range indexDesc.Fields {
-		indexResp.Fields = append(indexResp.Fields, fieldResponse{
-			Name:      field.Name,
-			Direction: field.Direction,
-		})
-	}
-
-	sendJSON(
-		req.Context(),
-		rw,
-		simpleDataResponse("index", indexResp),
-		http.StatusOK,
-	)
-}
-
-func dropIndexHandler(rw http.ResponseWriter, req *http.Request) {
-	db, err := dbFromContext(req.Context())
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	err = req.ParseForm()
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	colNameArg := req.Form.Get("collection")
-	indexNameArg := req.Form.Get("name")
-
-	col, err := db.GetCollectionByName(req.Context(), colNameArg)
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	err = col.DropIndex(req.Context(), indexNameArg)
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	sendJSON(
-		req.Context(),
-		rw,
-		simpleDataResponse("result", "success"),
-		http.StatusOK,
-	)
-}
-
-func listIndexHandler(rw http.ResponseWriter, req *http.Request) {
-	db, err := dbFromContext(req.Context())
-	if err != nil {
-		handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-		return
-	}
-
-	queryParams := req.URL.Query()
-	collectionParam := queryParams.Get("collection")
-
-	type fieldResponse struct {
-		Name      string                `json:"name"`
-		Direction client.IndexDirection `json:"direction"`
-	}
-
-	type indexResponse struct {
-		Name   string          `json:"name"`
-		ID     uint32          `json:"id"`
-		Fields []fieldResponse `json:"fields"`
-	}
-
-	convertToRep := func(indexDesc client.IndexDescription) indexResponse {
-		indexResp := indexResponse{
-			Name: indexDesc.Name,
-			ID:   indexDesc.ID,
-		}
-
-		for _, field := range indexDesc.Fields {
-			indexResp.Fields = append(indexResp.Fields, fieldResponse{
-				Name:      field.Name,
-				Direction: field.Direction,
-			})
-		}
-		return indexResp
-	}
-
-	if collectionParam == "" {
-		indexesPerCol, err := db.GetAllIndexes(req.Context())
-		if err != nil {
-			handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-			return
-		}
-		type collectionIndexes struct {
-			Collections map[client.CollectionName][]indexResponse `json:"collections"`
-		}
-		var resp collectionIndexes
-		resp.Collections = make(map[client.CollectionName][]indexResponse)
-		for colName, indexes := range indexesPerCol {
-			for _, index := range indexes {
-				resp.Collections[colName] = append(resp.Collections[colName], convertToRep(index))
-			}
-		}
-		sendJSON(
-			req.Context(),
-			rw,
-			struct {
-				Data collectionIndexes `json:"data"`
-			}{Data: resp},
-			http.StatusOK,
-		)
-	} else {
-		col, err := db.GetCollectionByName(req.Context(), collectionParam)
-		if err != nil {
-			handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-			return
-		}
-		indexes, err := col.GetIndexes(req.Context())
-		if err != nil {
-			handleErr(req.Context(), rw, err, http.StatusInternalServerError)
-			return
-		}
-		var resp []indexResponse
-		for _, index := range indexes {
-			resp = append(resp, convertToRep(index))
-		}
-		sendJSON(
-			req.Context(),
-			rw,
-			simpleDataResponse("indexes", resp),
-			http.StatusOK,
-		)
-	}
 }
 
 func getBlockHandler(rw http.ResponseWriter, req *http.Request) {
