+++ cli/index.go
@@ -17,7 +17,7 @@ import (
 func MakeIndexCommand() *cobra.Command {
 	var cmd = &cobra.Command{
 		Use:   "index",
-		Short: "Mange collections' indexes of a running DefraDB instance",
+		Short: "Manage collections' indexes of a running DefraDB instance",
 		Long:  `Manage (create, drop, or list) collection indexes on a DefraDB node.`,
 	}
